+++ src/app/(home)/product/page.js
@@ -1,12 +1,10 @@
 import SchedulingFeatures from '@/Product/SchedulingFeatures';
 import UserDocs from '@/Product/UserDocs';
-import WorldWideUsers from '@/Product/WorldWideUsers';
 
 const Product = () => {
     return (
       <>
         <SchedulingFeatures></SchedulingFeatures>
-        <WorldWideUsers></WorldWideUsers>
         <UserDocs></UserDocs>
       </>
     );
